Finish delete a product

+++ api-server/routes/product.route.js
@@ -2,6 +2,7 @@ const express = require('express');
 
 const Product = require('../models/Product');
 const productRoutes = express.Router();
+const { ObjectID } = require('mongodb');
 
 productRoutes.route('/').get((req, res) => {
     Product.find((err, products) => {
@@ -14,7 +15,7 @@ productRoutes.route('/').get((req, res) => {
 });
 
 productRoutes.route('/checkExist/:title').get((req, res) => {
-    // console.log('ddd');
+    console.log('ddd');
     const title = req.params.title;
     Product.findOne({title}).then(product => {
         if(product) {
@@ -36,14 +37,36 @@ productRoutes.route('/add').post((req, res) => {
         })
 });
 
-productRoutes.route('/delete/:id').get((req,res) => {
+productRoutes.route('/delete/:id').delete((req,res) => {
     let id = req.params.id;
+    if (!ObjectID.isValid(id)) {
+        return res.status(404).send('ID is invalid');
+    }
     console.log(id);
     Product.findOneAndDelete({_id: id})
         .then(product => {
             console.log(product);
-            res.status(200).json({rs: 'Delete success'});
+            if(product){
+                res.status(200).json({result: `Product ${product.title} is deleted succesfully`});
+            }else {
+                res.status(404).send('Cannot found product');
+            }
+        })
+        .catch(err => {
+            res.status(400).send('Cannot delete ', err);
         })
 });
+
+// productRoutes.route('/getInfo/:id').get((req, res) => {
+//     console.log('ddd');
+//     const id = req.params.id;
+//     Product.findOne({_id: id}).then(product => {
+//         if(product) {
+//             res.json(product);
+//         }else {
+//             res.status(200).send();
+//         }
+//     })
+// });
 
 module.exports = productRoutes;
